requirements.txt
absl-py==1.4.0
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beartype==0.14.0
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
datasets==2.12.0
decorator==5.1.1
diffusers==0.16.1
dill==0.3.6
einops==0.6.1
ema-pytorch==0.2.3
executing==1.2.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.14.1
idna @ file:///croot/idna_1666125576474/work
imagen-pytorch==1.25.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
kornia==0.6.12
lightning-utilities==0.8.0
lion-pytorch==0.0.7
lit==16.0.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
protobuf==4.21.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
Pygments==2.15.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytorch-warmup==0.1.1
pytz==2023.3
PyYAML==6.0
regex==2023.5.4
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
seaborn==0.12.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@567662b0fd33ca4b022d94d3b8de896628cd32dd
sentencepiece==0.1.99
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
thop==0.1.1.post2209072238
tokenizers==0.13.3
torch==2.0.1
torchaudio==0.12.0
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1680254681959/work
wcwidth==0.2.6
Werkzeug==2.3.4
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
